move util function and remove old analysis page and util
